requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
coverage==5.5
coveralls==1.11.1
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.1.7
django-allauth==0.54.0
django-countries==7.5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
drf-yasg==1.21.7
flake8==3.8.4
idna==3.4
inflection==0.5.1
mccabe==0.6.1
oauthlib==3.2.2
packaging==23.1
Pillow==10.0.0
pycodestyle==2.6.0
pycparser==2.21
pyflakes==2.2.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
responses==0.12.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
unittest-xml-reporting==3.0.4
uritemplate==4.1.1
urllib3==2.0.4
